Flattening of subfolders

+++ flatten_dataset.py
@@ -4,8 +4,7 @@ from pathlib import Path
 from tqdm import tqdm
 
 
-def flatten_dataset(source_dir, flattened_dir="data/cinc2020_flattened"):
-    # Create flattened directory if it does not exist
+def flatten_dataset(source_dir, flattened_dir):
     if not os.path.exists(flattened_dir):
         os.makedirs(flattened_dir)
 
@@ -16,17 +15,14 @@ def flatten_dataset(source_dir, flattened_dir="data/cinc2020_flattened"):
 
     # Walk through the source directory and list all .hea and .mat files
     for dirpath, dirnames, filenames in os.walk(source_dir):
-        # Check if this directory is a g* directory
-        if Path(dirpath).name.startswith("g"):
-            # Add each .hea and .mat file to the list
-            for filename in filenames:
-                if filename.endswith(".hea") or filename.endswith(".mat"):
-                    files_to_move.append(
-                        (
-                            os.path.join(dirpath, filename),
-                            os.path.join(flattened_dir, filename),
-                        )
+        for filename in filenames:
+            if filename.endswith(".hea") or filename.endswith(".mat"):
+                files_to_move.append(
+                    (
+                        os.path.join(dirpath, filename),
+                        os.path.join(flattened_dir, filename),
                     )
+                )
 
     # Move the files with a progress bar
     for src_file, dst_file in tqdm(
@@ -35,7 +31,10 @@ def flatten_dataset(source_dir, flattened_dir="data/cinc2020_flattened"):
         src_file = os.path.join(current_dir, str(src_file))
         dst_file = os.path.join(current_dir, (dst_file))
 
-        os.symlink(str(src_file), str(dst_file))
+        if not os.path.exists(dst_file):  # Check if file already exists
+            os.symlink(str(src_file), str(dst_file))
+        else:
+            print(f"File {dst_file} already exists. Skipping.")
 
     print(f"Flattened dataset is available at: {flattened_dir}")
 
@@ -44,7 +43,7 @@ def flatten_dataset(source_dir, flattened_dir="data/cinc2020_flattened"):
 if __name__ == "__main__":
     import sys
 
-    if len(sys.argv) != 2:
+    if len(sys.argv) != 3:
         print("Usage: python flatten_dataset.py <path_to_cinc2020_directory>")
     else:
-        flatten_dataset(sys.argv[1])
+        flatten_dataset(sys.argv[1], sys.argv[2])
